Low Altitude bonus
For your consideration. I created a low Alt Bonus. I use this on the NV map, the course ends at Nellis. if the player flies under the sunshades they get a bonus of 5 secons for each gate.

BonusGateHeight = <global height of the bonus gates in meters>
BonusGates = <list of gate numbers for low alt bonus>

This works like the Horizontal Gates, you add and array of gates to be allow the bonus to, then the height of the bonus alt. In my case 10m. or less through the gate and you get a 5 second bonus